requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

opencv-python==4.8.1.78
numpy==1.26.4
pillow==10.1.0

tensorflow==2.17.1
torch==2.2.2
torchvision==0.17.2

scikit-learn==1.4.2
pandas==2.2.2

python-dotenv==1.0.1
requests==2.31.0
mangum